cpuidle: Eliminate significant overhead from pm_qos latency notifier
Currently, the latency notifier kicks all CPUs out of idle all of the
time, even when the latency requirements are relaxed. Additionally, the
call to smp_call_function_many is set to wait by default, which causes
it to block for a while and waste a lot of CPU time since it is called
with preemption disabled.

Eliminate the mountain of overhead from the latency notifier by making
the smp_call_function_many non-blocking and only invoking it when the
CPU's latency requirements are tightened.

Additionally, remove the check for a mask in the v parameter since the
pm_qos framework does not pass anything for it.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Anush02198 <Anush.4376@gmail.com>